tests: move TemporaryDirectoryTestCase to util module
